feat(oc): enable name definition for deployment
Deployment were named the same as the namespace.

We often have more than one deployment per namespace so we want to be
able to declare its name.

We keep the namespace as default for the moment but this could be
improved to create something more unique (based on the container name?).

Maybe having default for a such an important resource does not make sense ?
We should base our decision based on real usage.

refs CICDITS-1058